update & delete elements / user & admin

+++ backend/controllers/post.js
@@ -30,9 +30,30 @@ exports.updatePost = (req, res, next) => {
         }`,
       }
     : { ...req.body };
-  Post.updateOne({ _id: req.params.id }, { ...postObject, _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Post modifié !" }))
-    .catch((error) => res.status(400).json({ error }));
+
+    delete postObject._userId;
+    Post.findOne({_id : req.params.id})
+    .then((post) =>{
+      if(req.auth.isAdmin === false){
+        if(post.userId != req.auth.userId){
+          res.satutus(401).json({message : "Non autorisé"});
+        }
+        else{
+        Post.updateOne({ _id: req.params.id }, { ...postObject, _id: req.params.id })
+      .then(() => res.status(200).json({ message: "Post modifié !" }))
+      .catch(error => res.status(401).json({ error }));
+        }
+      }
+      else{
+        Post.updateOne({ _id: req.params.id }, { ...postObject, _id: req.params.id })
+      .then(() => res.status(200).json({ message: "Post modifié !" }))
+      .catch(error => res.status(401).json({ error }));
+        }
+    })
+    .catch((error) => {
+      res.status(400).json({error});
+    })
+  
 };
 
 exports.deletePost = (req, res, next) => {
@@ -43,19 +64,29 @@ exports.deletePost = (req, res, next) => {
           error: new Error('Publication non disponible!')
         });
       }
+     if (req.auth.isAdmin === false){
       if (post.userId !== req.auth.userId) {
         res.status(400).json({
           error: new Error('Requête non autorisée!')
         });
       }
-     
-     else{const filename = post.imgUrl.split("/images/")[1];
+      else{const filename = post.imgUrl.split("/images/")[1];
       fs.unlink(`images/${filename}`, () => {
         Post.deleteOne({ _id: req.params.id })
           .then(() => res.status(200).json({ message: "Post supprimé !" }))
           .catch((error) => res.status(400).json({ error }));
       });
-}} )
+}
+    }
+    else{const filename = post.imgUrl.split("/images/")[1];
+      fs.unlink(`images/${filename}`, () => {
+        Post.deleteOne({ _id: req.params.id })
+          .then(() => res.status(200).json({ message: "Post supprimé !" }))
+          .catch((error) => res.status(400).json({ error }));
+      });
+}
+     
+     } )
     .catch((error) => res.status(500).json({ error }));
 };
